refactor: reponseDto 클래스 통합 (#25)
UserCommentsResponse.java와 UserPostsResponse.java의 내용을 UserCommunityResponse.java 클래스에 통합시켰다.
