+++ src/main/java/com/danthis/backend/api/user/UserController.java
@@ -3,9 +3,9 @@ package com.danthis.backend.api.user;
 import com.danthis.backend.api.ApiResponse;
 import com.danthis.backend.api.user.request.UserUpdateRequest;
 import com.danthis.backend.application.user.UserService;
+import com.danthis.backend.application.user.response.UserCommunityResponse.UserCommentsResponse;
+import com.danthis.backend.application.user.response.UserCommunityResponse.UserPostsResponse;
 import com.danthis.backend.application.user.response.UserInfoResponse;
-import com.danthis.backend.application.user.response.UserPostsResponse;
-import com.danthis.backend.application.user.response.UserCommentsResponse;
 import com.danthis.backend.common.security.aop.AssignCurrentUserInfo;
 import com.danthis.backend.common.security.aop.CurrentUserInfo;
 import io.swagger.v3.oas.annotations.Operation;
