--- src/main/java/com/danthis/backend/application/user/response/UserCommentsResponse.java
@@ -1,30 +1,0 @@
-package com.danthis.backend.application.user.response;
-
-import java.util.List;
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-@Builder
-public class UserCommentsResponse {
-
-  private List<CommentDto> comments;
-  private PaginationDto pagination;
-
-  @Getter
-  @Builder
-  public static class CommentDto {
-
-    private Long id;
-    private Long postId;
-    private String content;
-  }
-
-  @Getter
-  @Builder
-  public static class PaginationDto {
-
-    private Integer currentPage;
-    private Integer totalPages;
-  }
-}
